+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 ### Internals
 * Migrated to the final `std::filesystem` implementation on Windows from the experimental one.
 * Exception class InvalidKey is replaced with KeyNotFound, KeyAlreadyUsed, ColumnNotFound and ColumnAlreadyExists
+* Calling `Table::create_object(ObjKey)` on a table with a primary key column is now an error.
 
 ----------------------------------------------